Fix log levels
So turns out setting the [log level](https://docs.mongodb.com/mongoid/master/tutorials/mongoid-configuration/index.html#logging) in the config file does nothing. In a Rails environment Mongoid takes it's log level from Rails, and there doesn't seem to be a way to set it to something different.

Therefore I have taken the view that the logs are currentl too full of Mongoid/MongoDb output to be useful, and the sheer amount being written is likely to have a negative impact on performance.

Therefore this change resets the config file and switches the log level for the app to 'info' from now on.