Remove rake email anonymiser task
https://eaflood.atlassian.net/browse/WC-525

We simply don't think attempting to do this against collections of the size found in production is feasible from the app.

Hence we have instead re-written this task in Javascript which can be called from the mongo shell directly on the database server, which we hope will mean we can complete this task in a much shorter time.

Hence we are deleting this from the app.