Add workflow states to conviction_sign_offs
https://eaflood.atlassian.net/browse/WC-483

The current implementation of the conviction checking workflow allows a flagged renewal to be checked, then approved or rejected. However, the actual process is more complex, and a renewal which does have matching convictions goes through an internal process before a final decision is made. To better represent this process, we have updated the engine to include workflow states for the conviction_sign_off.

This commit updates the back office convictions process to use those states.