Display possible matches on index
After further thinking, I'm not convinced there is a user need for a page listing any renewal which was ever flagged for a conviction check regardless of the current status of that check. So instead we display the possible matches on the index and remove the old possible matches page.