List oldest renewals first on convictions dashboard
https://eaflood.atlassian.net/browse/WC-483

NCCC requested that the least recently modified renewals be listed first, as this matches the workflow they want to follow and lets them focus on renewals which have been waiting the longest. This was a minor change to the controller but led to a larger refactor of the tests, which has hopefully made them more readable.