Allow finance admin users to add 'stuck' WorldPay payments
https://eaflood.atlassian.net/browse/WC-503

If an external user makes a payment on WorldPay, but WorldPay fails to redirect them back to the app afterwards, then payment is taken but the app does not know about it, and leaves them in the `worldpay_form` state.

Finance picks up these 'missed' payments and records them, but isn't able to do this until the application has been submitted. This means that an agency user first has to move the application to a submitted state, and then a finance user adds the payment to complete the renewal.

To speed up this process, this change allows finance admin users to add a missed WorldPay payment once the renewal is in the `worldpay_form` state. Once the payment is added, the transient_registration should then progress to the next stage (either `renewal_received` or `renewal_complete` depending on other checks) as if the app received record of the payment directly from WorldPay.