Transition to submitted state after adding payment
When a missed WorldPay payment is successfully recorded, if the renewal was in the `worldpay_form` state, transition into a submitted state.

If the renewal can then be completed, it will complete. If further work is needed and the state is therefore `renewal_received_form`, the state change will still trigger the notification email so the user is aware their application has been received.