Refactor logic around can renew/complete checks
https://eaflood.atlassian.net/browse/WC-533

WC-533 highlighted that we were doing checks for whether something can be renewed in 2 different places, in 2 different ways. We agreed to move the checks to one place, but additional review has highlighted that we're using our current checks for mixed purposes, and as such are not being very efficient.

[Waste carriers renewals PR 307](https://github.com/DEFRA/waste-carriers-renewals/pull/307) did most of the refactoring, but we have also had to make changes to the back office, specifically to support how a renewal gets completed.

This covers the changes needed to support the new logic.

---

🎉 WIP 🎉