Fix broken tests due to date changes
Now the ExpiryCheckService in Waste Carriers Renewals handles calculating when the expiry date will be, it takes into account GMT/BST and as such strips the time of the return value.

So any test previously that matched both date and time started failing. The time is ignored in all logic based on `expires_on` so we can do the same in our tests. We just need to ensure the date is correct.

Hence these changes update tests that the changes had broken.