Correct invalid assumption with existing states
Had completely overlooked that the existing statues were not mutually exclusive e.g. if a renewal had outstanding conviction check and payment, then we would show both statuses, not just one of them.

So this change rectifies that, and uses a local variable to determine is a status has been set (to avoid calling both methods again).