Fix Airbrake for the dev environment
This fixes the issue where 500 error stack traces are swallowed up by the Airbrake module. When `ENV["USE_AIRBRAKE"]` is not `"true"` (as would be the case for the dev environment), this change successfully disables Airbrake.

Using `Airbrake.add_filter(&:ignore!)` is the recommended approach for Airbrake 5. The previous implementation would have worked in Airbrake 4.