Add privacy policy page for renewals
From https://eaflood.atlassian.net/browse/RUBY-302

Add verbal privacy policy screen on the backoffice.
The page's button "continue" will redirect to the beginning of the renewal journey for the given registration.
The front office dashboard renewal link will be changed to point to this page instead of directly the start of the journey.
This is the same strategy used for the verbal privacy policy on WEX.