Include registrations in dashboard search
https://eaflood.atlassian.net/browse/RUBY-704

The back office dashboard search was originally designed to only look at renewals - records in the TransientRegistrations table. We want to look at the Registrations table as well so we can see completed registrations and new applications in progress.

This will require updating our search service, as well as changing the design of the dashboard to be able to handle both types of object and show the correct information and actions for them.