Display correct number of results on dashboard
https://eaflood.atlassian.net/browse/RUBY-696

This fixes a bug found in QA. The result count in the table caption was showing the number on the page, not the total number.

We also will only return a maximum of 100 registrations and 100 transient_registrations to avoid overloading the service if someone searches for a common term and Kaminari then has to paginate an immense array.