Remove conviction details link in dashboard
NCCC have confirmed they would prefer conviction checks to only be managed from the conviction checks dashboard we built them as part of renewals.

So we have already removed links to a registrations conviction details from the view details screen. But we just spotted we still show a link in the dashboard search results.

So this is a minor change to the view to remove the link.