Create presenter for RenewingRegistration details page
This PR creates a presenter for RenewingRegistration, as we have done for Registration in details page.
It also includes a BaseRegistrationPresenter, which will take care of implementing shared functionality between the views.
Next step is adding partials for the different sections so that we can share code in the views which are incredibly long.