Update conviction dashboard to include registrations
https://eaflood.atlassian.net/browse/RUBY-791

Currently, the screen only handles submitted renewals which are awaiting approval.

In the old backend, there is no such screen. Users have to use BOXI to identify those with outstanding checks, then use the dashboard search to find and process them.

As part of the migration to the back-office, NCCC has asked that this screen be the only mechanism for accessing convictions checks in order to help streamline the process. So this PR updates the dashboard to include registrations which require approval as well.