Add controller, view and route
This is based off the registration transfer process, as that is the other place where we directly modify a registration rather than a transient_registration.

Lots of duplication in here, but given that this is temporary I think it's better to leave it this way so it's easy to rip out later.