Use BaseRegistrationForm for reg conviction forms
We can't inherit from BaseForm for this any more as it assumes there will be a token, something which registration objects don't have. So a hacky fix is to just make a very similar BaseRegistrationForm without that.

This could be refactored a lot more to make it more elegant, but these forms are temporary anyway, so I don't think it is worth the effort.