Activate and deactivate back office users
https://eaflood.atlassian.net/browse/RUBY-762

Super users should have the ability to activate and deactivate accounts of their group (eg agency supers can deactivate agency accounts, but not finance accounts).

If a user logs into a deactivated account, they should see a message telling them the account is disabled. They should not be able to access any information or make any changes. Active accounts can access and use the system as normal.

This means we can deal with revoking access to the system without entirely deleting all record of the user (which will be important for version control, etc later on).