Isolate finance details resource
We have realised that if we use the `finance_details._id` we are able to build a controller that does not depends on either parent. The controller contains a little hack which uses the same id to fetch the correct object by querying the two possible collections.
This refactoring have lead to cleaner code, remove of a lot of duplication and less need of helpers.