Add refund request to worldpay
From: https://eaflood.atlassian.net/browse/RUBY-821

Add code to send a refund request to Worldpay. This code only mimics what the old service is doing now. There is no check on wether the refund was successful or not.
Looking at the documentation, specifically here:  https://beta.developer.worldpay.com/docs/wpg/manage/modificationrequests#refund, in order to check the status of a refund we have to make a second call to the api and ask for the updated order details. This is then for another ticket.