Calculate the correct amount for refund
From: https://eaflood.atlassian.net/browse/RUBY-811

We have discovered that the refund should only trigger for outstanding amounts in registrations. This fixes the refund services so that the correct amount is refunded. In case the payment selected for the refund has an amount that is less than the overpaid balance, the refund amount will be set to the payment amount.