Change back office user role
https://eaflood.atlassian.net/browse/RUBY-761

It should be possible for super users to change the roles of other back office users.

This is dependent on the user group - for example, an agency super can only change the role of other agency users. The roles which are available to choose are also limited by the super user's group - so a finance super can only select finance roles.