Add form to role change page
This makes use of a partial to list the user roles, as we'll need to list the roles on the invite form as well.

The form is prepopulated with the user's current role, but we also display it as text in case the form is updated and a reminder is then required.