Remove excessive password validation
Devise handles password length validation for us with :validatable, so we don't need to add our own validation here.

We also don't need to attempt to validate the format when no password is provided as Devise should handle this.

This matches what we did on WEX and the advice from the Devise wiki: https://github.com/plataformatec/devise/wiki/How-To:-Set-up-simple-password-complexity-requirements