Do not show link to finance page if there is no finance details
From: https://eaflood.atlassian.net/browse/RUBY-808

QA have reported an error thrown by the page being visited from a transient registration that was still in progress. This was due to the fact that the page expect the presence of a `finance_details` object in the registration, but in case of in-progress registration there is none. Hence we shouldn't show the link there.