Fix wonky validation on user invite form
https://eaflood.atlassian.net/browse/RUBY-760

When we POST the invite form, there is an extra check to make sure that the user submitting the form has invited the new user to a group which they have permission for. (Eg a finance super has not invited an agency user).

This check was a little overzealous and was causing the other validation errors to not display correctly. Now it will only fire if the user has submitted a role and the role is invalid. We still don't display a helpful error but since this will only happen if the user is tinkering with the HTML of the form or trying to submit their own custom POST requests, I am not so worried about being helpful.

Also spotted that the link to email errors did not work and the error was not displayed next to the field. This PR fixes that as well.