Add 'developer' user role
https://eaflood.atlassian.net/browse/RUBY-772

We want to be able to update our conviction data from within the app. We'll need a new type of user group for this, called 'developer'.

Developer users will have the same permissions as agency_users to start out with. So we'll just consider them a part of the agency user group.