Add support for mocking external services
We have created a new engine in the DefraRuby namespace called [defra_ruby_mocks](https://github.com/DEFRA/defra-ruby-mocks).

Going forward this will allow us to mock external services like Companies House, and in the future Worldpay.

We only really need to do this when running performance tests. We want to put as much of the service under load as possible, but at the same time we don't want to get into trouble for overloading these external services, or worst case get accused of a DDOS attack.

So by providing an alternate end point that can accept the same request, and return what the service needs, we can replace them when we need to.

That endpoint is our existing apps. We have chosen to do it in this way to avoid the additional complexity of managing and maintaining another service across all our environments (if deployment was simple for us it's not the way we would have gone!)

So we'll be mounting the engine into both our [Front-office](https://github.com/DEFRA/waste-carriers-front-office) and [Back-office](https://github.com/DEFRA/waste-carriers-back-office) apps. We need to do this so each can treat the other as the mock endpoint (Ruby being single threaded means the rails apps can't call themselves)

This change adds the mocks engine to this app, including the initializer, which for now will give us the ability to mock companies house.

It also updates the [waste-carriers-engine](https://github.com/DEFRA/waste-carriers-engine) config. If mocking is enabled we want to tell the engine to use a different url for companies house lookups to allow us to override what the Companies house url is.

As both apps need to configure themselves to look at the other in an environment, we can't just update the `WCRS_COMPANIES_HOUSE_URL` as this is shared by everything.

Hence we have added logic to check if mocks are enabled. If they are, we refer to a different env var to get the URL to use. When mocks are not enabled, everything remains as it currently is.