Set up ConvictionImportService
https://eaflood.atlassian.net/browse/RUBY-772

Currently our convictions data is updated using the old waste-carriers-service in Java. We want to decomission that so we need a way for our new codebase to update the data.

This PR sets up a new ConvictionImportService which is designed to take raw data and update our list of convictions in the database.