Destroy old convictions data before adding new records
We want to avoid deleting the old convictions data and then raising an error, leaving the collection empty.

Ideally we would do all of this in a transaction to avoid that problem. But Mongoid 5 doesn't have them! 6 does though.

The best thing we can do is just hold off on making any database changes until we are as sure as possible that the new data is valid.