Fix bug in BOXI export
From: https://eaflood.atlassian.net/browse/RUBY-738

This fixes a bug in the generation of sign_offs files.
It also enhance the test so that we are sure we are at least testing a base scenario for insertion in all files. The curent code was never actually testing that a sign_off object passed through the feature without errors, we only had unit test coverage for it.