Set a worldpay response type
From: https://eaflood.atlassian.net/browse/RUBY-811

This adds the code that sets a worldpay status. In the old code, this is just arbitrarly set to `AUTHORISED`. This is used when showing that a payment has already been refunded, and as we want to show the user the same exact info as per old system, we need this value to be set.