Fix broken 'Finished' link on renewal end pages
https://eaflood.atlassian.net/browse/RUBY-859

We spotted that the 'Finished' button on the 'Renewal complete' and 'Renewal received' buttons was not behaving properly.

This PR overrides the new default in the engine to direct back office users to the 'view details' page. If the renewing registration no longer exists, they will be redirected to the page for the registration instead.