View registration certificate in back office
https://eaflood.atlassian.net/browse/RUBY-834

We already email out certificates from the back office with completed renewals.

However, the old backend also has a feature where you can view a certificate on-demand through the back office. The PDF generation is already in place in the engine. We just need to load it from a route.