Add controller and route to view certificate in app
This is basically copied from what we do to render PDFs in the WEX back office - in that case, confirmation letters and renewal letters.

This is currently rendering as HTML no matter what for some reason. Still working on fixing that. But committing for now before my brain turns to mush.