Delete one off address fix rake task
https://eaflood.atlassian.net/browse/WC-500

WC-500 was a bug found with addresses in renewals and renewed registrations found after the new renewals system went live in Spet 2018.

We fixed the bug, but part of the solution was also correcting the existing records so we added a rake task to do just that.

It was only ever intended to be a one-off; once run run the existing records would be fixed and changes made to the code would ensure new ones were as expected.

So this change is just about deleting more than 200 lines of code we really don't need anymore 😄