Remove unused passenger rake tasks
In our AWS environments we use passenger as our app server, and have config and systemd scripts to do that.

Initially we build our local Vagrant environment to also use passenger, and even had it defaulted when running it directly on our local host machines.

But to put it bluntly, we found it to be a pain in the 🍑!

So it was eventually dropped as the default, then from our vagrant machines. We still use it in production where we have found it useful, we just want it out of the way when working locally.

At the time we added it to the project we also added some rake tasks, because you know, they _might_ come in handy!? Turns out we never use these them.

So as part of keeping things tidy, and not keeping code around just for the sake of it might one day be used, this change deletes the unused rake tasks.