Remove expired panel from renewing registration details page
We have been reported a bug for which requesting the details page of a Renewing registration created AFTER the expire date of a registration was returning a 500 error.
Further debugging showed that the problem is in the fact that the Renewing Registration does not implement methods used in that partial to decide what info to show.
We have realised that the panel should actually never show for renewing registrations. Hence this removes it entirely.