Add view payments ability
Currently, abilities for managing payments and reversale are bounded to the type of payment the user can deal with. Given that, in order to being able to block a user from viewving intermediary screens and links to those functionality me and Iris have decided to add a `view_payments` ability.