Redirect to correct resource page after renewal is completed
From: https://eaflood.atlassian.net/browse/RUBY-895

We have realised that once a renewal have been completed we have the need to redirect the user to the correct resource page, as the transient resource will be deleted at that point.
This contains the refactoring and fix necessary to make that happen.