Remove `renewal` from payment pages
From: https://eaflood.atlassian.net/browse/RUBY-826

It was reported that the current text on payment pages was referring to renewals even when the payment was taken on registrations.
In order to avoid adding extra complication to the code base, I took the liberty of adjusting the text so that it will apply to either a renewal or a registration.