Fix renewal bug on charge adjust
From: https://eaflood.atlassian.net/browse/RUBY-870

We have realised that by running the service for charge adjust in the controller, we are in fact updating the finance details after the renewal completion service had the chance of running.
With this we have refactored the forms to run the service and the ResourceFormController to make it possible to inject a current_user for the service to use.

It also fixes some messages on the translations which where spelled incorrectly.