Add hint to all text area
From: https://eaflood.atlassian.net/browse/RUBY-890

This adds an hint to all of our text areas missing one to let the user know about the maximum number of characters that can be inserted into the text area.

We have also agreed to fix all validations for text area to have a maximum of 500 characters all across the project, hence this PR also fixes the validation of payment forms which was set to 250.