Hide edit functionality behind a feature toggle
https://eaflood.atlassian.net/browse/RUBY-926

We have implemented feature toggle functionality in the [waste-carriers-engine](https://github.com/DEFRA/waste-carriers-engine/pull/697) which gives us a state we can check in one place, and control through config.

This is to allow us to hide functionality that is not yet ready for release in production, but access it in our non-production environments.

One such feature that is still being developed is the edit registration feature. So this change takes advantage of the new feature toggles, and hides the edit links if the feature is disable.