Add config file
To allow us to switch it on or off without changing the code use an env var to set the active state.

We default to `true` if the env var isn't there as that is our default state, and it means we don't have to monkey about with the tests. But when we release we'll have the env var set to `false`.