Don't set the feature on as the default
Again another great suggestion by @irisfaraway. To be sure we don't accidently have this on in production because we failed to set an env var, we want the default to be `false`.

If no env var is set the feature toggle classes will return `false` so removing the `|| true` does the trick (thanks to @cintamani for that one).